requirements.txt
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.6
django-admin-list-filter-dropdown==1.0.3
django-admin-rangefilter==0.5.1
django-geojson==2.12.0
django-grappelli==2.13.1
django-import-export==1.2.0
django-leaflet==0.25.0
django-summernote==0.8.11.4
et-xmlfile==1.0.1
Faker==2.0.3
jdcal==1.4.1
jsonfield==2.0.2
MarkupPy==1.14
odfpy==1.4.0
openpyxl==3.0.0
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.1.2
six==1.12.0
sqlparse==0.3.0
tablib==0.14.0
text-unidecode==1.3
xlrd==1.2.0
xlwt==1.3.0
